account create page + raw numpad

+++ lib/models.dart
@@ -1,9 +1,9 @@
-import 'package:fb/repository.dart';
+import 'package:fb/db/category.dart';
 
 class CategoryStat {
   final Category category;
-  final int left;
-  final int total;
+  final double left;
+  final double total;
   final String currency;
 
   CategoryStat(this.category, this.left, this.total, this.currency);
